Refactor SSF Receiver support
- Move issuer/audience validation into DefaultSsfEventProcessor
- Simplify SsfPushDeliveryResource
- Validate SecurityEventToken type
