+++ services/src/main/java/org/keycloak/protocol/ssf/endpoint/SsfPushDeliveryResource.java
@@ -1,7 +1,5 @@
 package org.keycloak.protocol.ssf.endpoint;
 
-import java.util.Set;
-
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.HeaderParam;
 import jakarta.ws.rs.POST;
@@ -11,7 +9,6 @@ import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.HttpHeaders;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
-import jakarta.ws.rs.core.UriInfo;
 
 import org.keycloak.models.KeycloakContext;
 import org.keycloak.models.KeycloakSession;
@@ -22,9 +19,7 @@ import org.keycloak.protocol.ssf.event.processor.SsfEventContext;
 import org.keycloak.protocol.ssf.receiver.SsfReceiver;
 import org.keycloak.protocol.ssf.receiver.registration.SsfReceiverRegistrationProviderFactory;
 import org.keycloak.protocol.ssf.receiver.spi.SsfReceiverProvider;
-import org.keycloak.services.Urls;
 import org.keycloak.services.resources.KeycloakOpenAPI;
-import org.keycloak.urls.UrlType;
 
 import org.eclipse.microprofile.openapi.annotations.Operation;
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
@@ -32,7 +27,7 @@ import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.jboss.logging.Logger;
 
-import static org.keycloak.utils.KeycloakSessionUtil.getKeycloakSession;
+import org.keycloak.utils.KeycloakSessionUtil;
 
 /**
  * SsfPushDeliveryResource implements the RFC 8935 Push-Based Security Event Token (SET) Delivery Using HTTP.
@@ -101,7 +96,7 @@ public class SsfPushDeliveryResource {
                                              @HeaderParam(HttpHeaders.CONTENT_TYPE) String contentType //
     ) {
 
-        KeycloakSession session = getKeycloakSession();
+        KeycloakSession session = KeycloakSessionUtil.getKeycloakSession();
         KeycloakContext context = session.getContext();
 
         SsfReceiver receiver = lookupReceiver(session, receiverAlias, context);
@@ -128,13 +123,9 @@ public class SsfPushDeliveryResource {
         }
 
         // Security Event Token is parsed and signature validated from here on
-        LOG.debugf("Ingesting valid security event token. realm=%s receiverAlias=%s jti=%s", realm.getName(), receiverAlias, securityEventToken.getId());
+        LOG.debugf("Ingesting security event token. realm=%s receiverAlias=%s jti=%s", realm.getName(), receiverAlias, securityEventToken.getId());
 
-        // Perform additional validations
-        checkIssuer(session, receiver, securityEventToken, securityEventToken.getIssuer());
-        checkAudience(session, receiver, securityEventToken, securityEventToken.getAudience());
-
-        // Security Event Token is valid
+        // Security Event Token parsed
         eventContext.setSecurityEventToken(securityEventToken);
 
         handleEvents(session, securityEventToken, eventContext);
@@ -165,15 +156,6 @@ public class SsfPushDeliveryResource {
         ssfReceiverProvider.processEvents(securityEventToken, eventContext);
     }
 
-    protected void checkIssuer(KeycloakSession session, SsfReceiver receiver, SecurityEventToken securityEventToken, String issuer) {
-
-        String expectedIssuer = receiver.getConfig() != null ? receiver.getConfig().getIssuer() : null;
-
-        if (!isValidIssuer(receiver, expectedIssuer, issuer)) {
-            throw SsfSetPushDeliveryFailureResponse.newFailureResponse(Response.Status.BAD_REQUEST, SsfSetPushDeliveryFailureResponse.ERROR_INVALID_ISSUER, "Invalid issuer");
-        }
-    }
-
     protected void checkPushAuthorizationToken(KeycloakSession session, SsfReceiver receiver, String receivedAuthHeader) {
 
         String expectedAuthHeader = receiver.getConfig() != null ? receiver.getConfig().getPushAuthorizationHeader() : null;
@@ -183,34 +165,6 @@ public class SsfPushDeliveryResource {
                 throw SsfSetPushDeliveryFailureResponse.newFailureResponse(Response.Status.BAD_REQUEST, SsfSetPushDeliveryFailureResponse.ERROR_AUTHENTICATION_FAILED, "Invalid push authorization header");
             }
         }
-    }
-
-    protected void checkAudience(KeycloakSession session, SsfReceiver receiver, SecurityEventToken securityEventToken, String[] audience) {
-
-        Set<String> expectedAudience = receiver.getConfig() != null && receiver.getConfig().getStreamAudience() != null ? receiver.getConfig().streamAudience() : null;
-
-        if (expectedAudience == null) {
-            // No expected audience configured for receiver, fallback to realm issuer is no audience is set
-            String fallbackAudience = getFallbackAudience(session);
-            expectedAudience = Set.of(fallbackAudience);
-        }
-
-        if (!isValidAudience(receiver, expectedAudience, audience)) {
-            throw SsfSetPushDeliveryFailureResponse.newFailureResponse(Response.Status.BAD_REQUEST, SsfSetPushDeliveryFailureResponse.ERROR_INVALID_AUDIENCE, "Invalid audience");
-        }
-    }
-
-    protected String getFallbackAudience(KeycloakSession session) {
-        UriInfo frontendUriInfo = session.getContext().getUri(UrlType.FRONTEND);
-        return Urls.realmIssuer(frontendUriInfo.getBaseUri(), session.getContext().getRealm().getName());
-    }
-
-    protected boolean isValidIssuer(SsfReceiver receiver, String expectedIssuer, String issuer) {
-        return expectedIssuer.equals(issuer);
-    }
-
-    protected boolean isValidAudience(SsfReceiver receiver, Set<String> expectedAudience, String[] audience) {
-        return expectedAudience.containsAll(Set.of(audience));
     }
 
     protected boolean isValidPushAuthorizationHeader(SsfReceiver receiver, String authHeader, String expectedAuthHeader) {
